requirements.txt
numpy==1.21.2
pandas==1.3.2
python-dateutil==2.8.2
python-igraph==0.9.6
pytz==2021.1
six==1.16.0
texttable==1.6.4
